requirements.txt
# Core dependencies
numpy>=1.24.0
pydantic>=2.0.0
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0

# Gradio (install last)
gradio>=4.0.0